requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.18.0
Flask==3.1.1
fonttools==4.59.0
frozenlist==1.6.2
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.7
loguru==0.7.3
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
nvidia-cublas-cu12==12.4.2.65
nvidia-cuda-cupti-cu12==12.4.99
nvidia-cuda-nvrtc-cu12==12.4.99
nvidia-cuda-runtime-cu12==12.4.99
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.0.44
nvidia-curand-cu12==10.3.5.119
nvidia-cusolver-cu12==11.6.0.99
nvidia-cusparse-cu12==12.3.0.142
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.4.99
openai==1.97.0
packaging==25.0
pandas==2.3.0
pillow==11.0.0
propcache==0.3.1
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rouge-chinese==1.0.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.12.2
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
text2vec==1.3.5
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchvision==0.19.0+cu124
tqdm==4.67.1
transformers==4.53.2
triton==3.0.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
volcengine-python-sdk==3.0.9
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
zhconv==1.4.3
zipp==3.23.0
